Support searching arbitrary headers
This is a preliminary implementation (no documentation yet) for
https://github.com/rc0/mairix/issues/13 that also fixes
https://github.com/vandry/mairix/issues/11. With this change, my
database indexing 57,000 emails grew from ~50 MB to ~230 MB, so indexing
minor headers is an opt-in feature that must be enabled by adding
"index_all_headers" to the mairix configuration file. I'm considering
making it possible to select which headers get indexed by explicitly
specifying headers that user does or does not want to be tracked.